requirements.txt
aiogram
aioschedule
python-dotenv
aiohttp
RPi.GPIO
picamera